backend/requirements.txt
# Quant AI Lab - Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==13.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.0
python-multipart==0.0.9
httpx==0.27.0
aiohttp==3.10.0

# Database
sqlalchemy==2.0.35
alembic==1.13.0
aiosqlite==0.20.0

# Data & ML
pandas==2.2.0
numpy==1.26.0
scipy==1.14.0
scikit-learn==1.5.0
xgboost==2.1.0
lightgbm==4.5.0
torch==2.4.0
vectorbt==0.26.2

# Technical Analysis
ta==0.11.0
pandas-ta==0.3.14b

# Broker APIs
smartapi-python==1.3.6
kiteconnect==5.0.1
# dhanhq==2.0.0
# fyers-apiv3==3.0.0

# LLM
openai==1.50.0
anthropic==0.34.0
ollama==0.3.0

# Communication
python-telegram-bot==21.5

# Redis
redis==5.1.0

# Vector DB
chromadb==0.5.0
sentence-transformers==3.1.0

# Utilities
pyyaml==6.0.2
python-dotenv==1.0.1
loguru==0.7.2
orjson==3.10.0
pillow==10.4.0

# Voice (for WhatsApp voice messages)
openai-whisper==20231117

# Packaging
pyinstaller==6.10.0
